+++ examples/nextjs/components/Web3Provider.tsx
