+++ examples/nextjs/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 
-import { Web3Provider } from '../components/web3';
+import { Web3Provider } from '../components/Web3Provider';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
